feat: Address 위경도 추가 및 매퍼 함수 분리
